Expand supported distance units, add initial weight units
This makes it so that feet and inches are supported for converting
distance. It also adds some initial weight units for converting between
pounds and ounces--more units to be added in the future.

This also ensures that the from and to units are able to be converted
and of the same type.

New errors are introduced to add clarity to invalid or incompatible unit
types being passed in.