Refactor to appease RuboCop, update config
This gets RuboCop passing so that it can be used with CI checks. This
is for Episode 04 of the screencast series. I think RuboCop has some
value in that it can catch errors before committing them and lead to
refactoring code to be more readable and abide to principals a group
agrees upon. But it's easy to take it too far and make it frustrating
for other contributors.

Anyway, this includes some code refactors to appease RuboCop, but no new
functionality.